more math unit tests and consolidation
